before check out cart will be checked for inventory in DB

+++ src/main/java/cova/rar/service/OrderService.java
@@ -1,6 +1,6 @@
 package cova.rar.service;
 
-import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 import java.sql.Date;
@@ -10,7 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import cova.rar.dao.OrderDao;
 import cova.rar.dao.ProductDao;
 import cova.rar.entities.Cart;
+import cova.rar.entities.Item;
 import cova.rar.entities.Order;
+import cova.rar.entities.Product;
 
 public class OrderService {
 
@@ -20,18 +22,31 @@ public class OrderService {
 	@Autowired
 	ProductDao productDao;
 
-	public void addCart(Cart cart, String userID) {
-
+	public Cart addCart(Cart cart, String userID) {
 		// call dao method to add items to item table and reference that to a single
 		// order id
+		boolean cartUpdated = false;
 		
 		Date date = new Date(Calendar.getInstance().getTime().getTime());
 		int orderID = orderDao.insertOrder(userID, date);
 		
-		orderDao.addItems(orderID, cart.getItems());
+		// check to see if inventory is greater or equal to cart qty
+		for (Item item : cart.getItems()) {
+			Product prodInDB = productDao.getProduct(item.getProduct().getId());
+			if (prodInDB.getInventory() < item.getQty()) {
+				cart.updateItemQty(item, prodInDB.getInventory());
+				cartUpdated = true;
+			}
+		}
 		
+		if (cartUpdated) {
+			return cart;
+		}
+		
+		orderDao.addItems(orderID, cart.getItems());
 		productDao.removeInventory(cart.getItems());
 		
+		return null;
 	}
 	
 	public List<Order> getOrderHistory(String userID) {
